requirements.txt
qdrant-client>=1.9.0
sentence-transformers>=2.7.0
huggingface-hub>=0.23.0
gradio>=4.36.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
datasets>=2.19.0
tqdm>=4.66.0
